+++ movieland-backend/src/main/java/com/phil/movieland/rest/request/ReservationRequest.java
@@ -5,14 +5,14 @@ import com.phil.movieland.data.entity.Seat;
 import java.util.List;
 
 public class ReservationRequest {
-    private long show_id;
+    private int show_id;
     private List<Seat> seats;
 
-    public long getShow_id() {
+    public int getShow_id() {
         return show_id;
     }
 
-    public void setShow_id(long show_id) {
+    public void setShow_id(int show_id) {
         this.show_id=show_id;
     }
